Add hello-world post
Use the main tag to create the main content canvas for all the posts.
Use article tag to begin a new post with hello-world as its id
attribute. Create two headers for both article title and date to have
separate stylings and positions in the article. Add section for the main
content of the post.