Add styling for main section and headers
Create styling for the main section to have a different background and
have rounded corners for it to be distinct from the larger body.

Create styling for both article and date headers in the main section
that allows each to be aligned together on the left, having the date
right below the article title.

Change font from Inconsolata to Monospace because not all computers have
Inconsolata.