Change multiple stylings in stylesheet
Move font-family to body instead of using the *. This is a performance
fix, to allow the cascading to only work on the body, and not to
everything. Change the font to a san-serif font family to be more
readable on a digital screen

Align header to be centered for symmetrical design. Make the size
smaller than previous to fit the page better.

Make the navigation bar's padding zero to account for the small padding
to the left that an unordered list creates. Make the list elements (li)
have no wrap so that each element of the navigation bar does not wrap
accordingly.

Make <br> and <p> have extra spacing underneath. Change <br>'s margin
bottom to compensate for a double line change. Add padding to <p> to
have an extra line before the next paragraph.

Add limitations to the main body for max and min width to respond to
whatever screen the user views the website on. Make max height automatic
to compensate for the change each time content is added.

Remove the margins and paddings of the article-header and add margins to
the top of the date-header to have a bit of spacing between the content
title and post-date.

Remove all margins in the main-content section to wrap into the article
margins instead.